+++ Integrations/ThreatConnect/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## [Unreleased]
-Added new commands.
+Added 8 new commands.
   - ***tc-get-groups***
   - ***tc-add-group-security-label***
   - ***tc-add-group-tag***
@@ -7,3 +7,4 @@ Added new commands.
   - ***tc-group-associate-indicator***
   - ***tc-get-events***
   - ***tc-add-group-attribute***
+  - ***tc-create-document-group***